Base64 Messages Should Not Decrypt
Base64 messages sometimes incorrectly decrypt to junk data. Doesn't
let the user read the message but it does make the GUI jump around a
bit. Lets add in a specific check before trying to decode or decrypt a
message to determine if it is a valid action. Note, this is only
relevent because I want to keep the original chat history for demo
purposes.