Backup And Load Chat History To Skynet
Every client should keep a Skynet backup of all their chat channels.
This way if the data is lost from GunDb computers that had the data
will continue to have the data. Also, re-anounce all data to the
GunDb network that was loadedin from the Skynet backup. This will help
the chat to recover from lost data.